requirements.txt
pyserial
pyubx2
termcolor